+++ examples/relay/CHANGELOG.md
@@ -1,5 +1,12 @@
 # relay-example
 
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - lex-gql@0.5.0
+
 ## 0.0.4
 
 ### Patch Changes